Extract Mapping classes to own module
