+++ beetsplug/filetote.py
@@ -5,7 +5,7 @@ from dataclasses import asdict, dataclass
 from sys import version_info
 from typing import Any, List, Optional, Tuple, Union
 
-from beets import config, dbcore, util
+from beets import config, util
 from beets.library import DefaultTemplateFunctions, Item
 from beets.plugins import BeetsPlugin
 from beets.ui import get_path_formats
@@ -13,71 +13,12 @@ from beets.util import MoveOperation
 from beets.util.functemplate import Template
 from mediafile import TYPES as BEETS_FILE_TYPES
 
+from .mapping_model import FiletoteMappingFormatted, FiletoteMappingModel
+
 if version_info >= (3, 8):
     from typing import Literal
 else:
     from typing_extensions import Literal  # type: ignore # pylint: disable=import-error
-
-
-class FiletoteMappingModel(dbcore.db.Model):
-    """Model for a FormattedFiletoteMapping."""
-
-    _fields = {
-        "artist": dbcore.types.STRING,
-        "albumartist": dbcore.types.STRING,
-        "album": dbcore.types.STRING,
-        "albumpath": dbcore.types.STRING,
-        "medianame_old": dbcore.types.STRING,
-        "medianame_new": dbcore.types.STRING,
-        "old_filename": dbcore.types.STRING,
-    }
-
-    def set(self, key: str, value: str) -> None:
-        """Get the formatted version of model[key] as string."""
-        return super().__setitem__(key, value)
-
-    @classmethod
-    def _getters(cls) -> dict:
-        """Returnblank for getter functions."""
-        return {}
-
-    def _template_funcs(self) -> dict:
-        """Return blank for template functions."""
-        return {}
-
-
-class FiletoteFormattedMapping(dbcore.db.FormattedMapping):
-    """
-    Formatted Mapping that does not replace path separators for certain keys
-    (e.g., albumpath).
-    """
-
-    ALL_KEYS: Literal["*"] = "*"
-
-    def __init__(
-        self,
-        model: FiletoteMappingModel,
-        included_keys: Union[Literal["*"], list] = ALL_KEYS,
-        for_path: bool = False,
-        whitelist_replace: Optional[List[str]] = None,
-    ):
-        super().__init__(model, included_keys, for_path)
-        if whitelist_replace is None:
-            whitelist_replace = []
-        self.whitelist_replace = whitelist_replace
-
-    def __getitem__(self, key: str) -> str:
-        """
-        Get the formatted version of model[key] as string. Any value
-        provided in the `whitelist_replace` list will not have the path
-        separator replaced.
-        """
-        if key in self.whitelist_replace:
-            value = self.model._type(key).format(self.model.get(key))
-            if isinstance(value, bytes):
-                value = value.decode("utf-8", "ignore")
-            return value
-        return super().__getitem__(key)
 
 
 @dataclass
@@ -168,6 +109,7 @@ class FiletotePlugin(BeetsPlugin):
         self.register_listener("cli_exit", self.process_events)
 
     def _get_filetote_path_formats(self, queries: List[str]) -> List[tuple]:
+        """Gets all `path` formats from beets and parses those set for Filetote."""
         path_formats = []
 
         for path_format in get_path_formats():
@@ -274,7 +216,7 @@ class FiletotePlugin(BeetsPlugin):
         file_name_no_ext = util.displayable_path(os.path.splitext(filename)[0])
         mapping.set("old_filename", file_name_no_ext)
 
-        mapping_formatted = FiletoteFormattedMapping(
+        mapping_formatted = FiletoteMappingFormatted(
             mapping, for_path=True, whitelist_replace=["albumpath"]
         )
